requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Celery (비동기 처리)
redis==5.0.1
celery==5.3.6
django-redis==5.4.0

# Authentication & Security
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
cryptography==42.0.2
requests==2.31.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.1

# API Documentation
drf-spectacular==0.27.1

# Code Quality & Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0
black==24.1.1
ruff==0.1.15
isort==5.13.2
pre-commit==3.6.0

# Monitoring & Logging
django-extensions==3.2.3

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
ipython==8.20.0
django-debug-toolbar==4.2.0
